server/pyproject.toml
[project]
name = "sweet-home-server"
version = "1.0.0"
description = "Backend for Sweet Home"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fpdf>=1.7.2",
    "pillow>=12.0.0",
    "python-barcode>=0.16.1",
    "qrcode>=8.2",
    "quart>=0.20.0",
    "quart-cors>=0.8.0",
    "quart-jwt-extended>=0.1.0",
    "setuptools>=80.9.0",
    "werkzeug>=3.1.3",
]
